+++ common/src/main/java/com/chefmooon/colourfulclocks/client/model/ColourfulClocksModels.java
@@ -1,0 +1,13 @@
+package com.chefmooon.colourfulclocks.client.model;
+
+// TODO : client init()?
+public class ColourfulClocksModels {
+    public static final String MINUTE_HAND_PATH = "item/%s_minute_hand";
+    public static final String HOUR_HAND_PATH = "item/%s_hour_hand";
+    public static final String MINUTE_HAND_SMALL_PATH = "item/%s_minute_hand_small";
+    public static final String HOUR_HAND_SMALL_PATH = "item/%s_hour_hand_small";
+    public static final String BORNHOLM_DIAL_PATH = "block/%s_bornholm_dial";
+    public static final String DIAL_SMALL_PATH = "block/%s_dial_small";
+    public static final String BORNHOLM_DOOR_PATH = "block/%s_bornholm_door";
+
+}