+++ fabric/src/main/java/com/chefmooon/colourfulclocks/client/event/fabric/ClientSetupEventsImpl.java
@@ -1,8 +1,12 @@
 package com.chefmooon.colourfulclocks.client.event.fabric;
 
+import com.chefmooon.colourfulclocks.client.model.ColourfulClocksModels;
 import com.chefmooon.colourfulclocks.client.renderer.fabric.BornholmMiddleBlockEntityRendererImpl;
 import com.chefmooon.colourfulclocks.client.renderer.fabric.BornholmTopBlockEntityRendererImpl;
 import com.chefmooon.colourfulclocks.client.renderer.fabric.MantelClockBlockEntityRendererImpl;
+import com.chefmooon.colourfulclocks.common.data.types.BornholmDoorTypes;
+import com.chefmooon.colourfulclocks.common.data.types.BornholmTopGlassTypes;
+import com.chefmooon.colourfulclocks.common.data.types.WoodTypes;
 import com.chefmooon.colourfulclocks.common.registry.fabric.ColourfulClocksBlockEntitiesImpl;
 import com.chefmooon.colourfulclocks.common.registry.fabric.ColourfulClocksBlocksImpl;
 import com.chefmooon.colourfulclocks.common.registry.fabric.ColourfulClocksItemsImpl;
@@ -17,6 +21,7 @@ import net.minecraft.world.level.block.Block;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
+@SuppressWarnings(("deprecation"))
 public class ClientSetupEventsImpl {
 
     public static void onRegisterRenderers() {
@@ -24,28 +29,41 @@ public class ClientSetupEventsImpl {
         BlockEntityRenderers.register(ColourfulClocksBlockEntitiesImpl.BORNHOLM_TOP_VARIANTS, BornholmTopBlockEntityRendererImpl::new);
         BlockEntityRenderers.register(ColourfulClocksBlockEntitiesImpl.MANTEL_CLOCK_VARIANTS, MantelClockBlockEntityRendererImpl::new);
 
-        BlockRenderLayerMap.INSTANCE.putBlocks(RenderType.translucent(),
+        BlockRenderLayerMap.INSTANCE.putBlocks(RenderType.cutout(),
                 ColourfulClocksBlocksImpl.BORNHOLM_TOP_VARIANTS.values().stream().map(Supplier::get).toArray(Block[]::new)
         );
-        BlockRenderLayerMap.INSTANCE.putBlocks(RenderType.translucent(),
+        BlockRenderLayerMap.INSTANCE.putBlocks(RenderType.cutout(),
                 ColourfulClocksBlocksImpl.BORNHOLM_MIDDLE_VARIANTS.values().stream().map(Supplier::get).toArray(Block[]::new)
         );
-        BlockRenderLayerMap.INSTANCE.putBlocks(RenderType.translucent(),
+        BlockRenderLayerMap.INSTANCE.putBlocks(RenderType.cutout(),
                 ColourfulClocksBlocksImpl.MANTEL_CLOCK_VARIANTS.values().stream().map(Supplier::get).toArray(Block[]::new)
         );
     }
 
     public static void onRegisterModels(Consumer<ResourceLocation> consumer) {
         ColourfulClocksItemsImpl.POCKET_WATCH_VARIANTS.forEach(((pocketWatchTypes, itemSupplier) -> {
-            ResourceLocation minuteHandLocation = TextUtil.res("item/%s_minute_hand".formatted(BuiltInRegistries.ITEM.getKey(itemSupplier.get()).getPath()));
+            ResourceLocation minuteHandLocation = TextUtil.res(ColourfulClocksModels.MINUTE_HAND_PATH.formatted(BuiltInRegistries.ITEM.getKey(itemSupplier.get()).getPath()));
             consumer.accept(minuteHandLocation);
-            ResourceLocation hourHandLocation = TextUtil.res("item/%s_hour_hand".formatted(BuiltInRegistries.ITEM.getKey(itemSupplier.get()).getPath()));
+            ResourceLocation hourHandLocation = TextUtil.res(ColourfulClocksModels.HOUR_HAND_PATH.formatted(BuiltInRegistries.ITEM.getKey(itemSupplier.get()).getPath()));
             consumer.accept(hourHandLocation);
 
-            ResourceLocation minuteHandSmallLocation = TextUtil.res("item/%s_minute_hand_small".formatted(BuiltInRegistries.ITEM.getKey(itemSupplier.get()).getPath()));
+            ResourceLocation minuteHandSmallLocation = TextUtil.res(ColourfulClocksModels.MINUTE_HAND_SMALL_PATH.formatted(BuiltInRegistries.ITEM.getKey(itemSupplier.get()).getPath()));
             consumer.accept(minuteHandSmallLocation);
-            ResourceLocation hourHandSmallLocation = TextUtil.res("item/%s_hour_hand_small".formatted(BuiltInRegistries.ITEM.getKey(itemSupplier.get()).getPath()));
+            ResourceLocation hourHandSmallLocation = TextUtil.res(ColourfulClocksModels.HOUR_HAND_SMALL_PATH.formatted(BuiltInRegistries.ITEM.getKey(itemSupplier.get()).getPath()));
             consumer.accept(hourHandSmallLocation);
         }));
+        for (BornholmTopGlassTypes type : BornholmTopGlassTypes.values()) {
+            ResourceLocation location = TextUtil.res(ColourfulClocksModels.BORNHOLM_DIAL_PATH.formatted(type.getName()));
+            consumer.accept(location);
+        }
+        for (BornholmTopGlassTypes type : BornholmTopGlassTypes.values()) {
+            ResourceLocation location = TextUtil.res(ColourfulClocksModels.DIAL_SMALL_PATH.formatted(type.getName()));
+            consumer.accept(location);
+        }
+        for (BornholmDoorTypes doorTypes : BornholmDoorTypes.values()) {
+            if (doorTypes == BornholmDoorTypes.BASE) continue; // Skip empty door type
+            ResourceLocation location = TextUtil.res(ColourfulClocksModels.BORNHOLM_DOOR_PATH.formatted(doorTypes.getName()));
+            consumer.accept(location);
+        }
     }
 }