+++ neoforge/src/main/java/com/chefmooon/colourfulclocks/client/renderer/neoforge/MantelClockBlockEntityRendererImpl.java
@@ -1,8 +1,10 @@
 package com.chefmooon.colourfulclocks.client.renderer.neoforge;
 
+import com.chefmooon.colourfulclocks.client.model.ColourfulClocksModels;
 import com.chefmooon.colourfulclocks.client.renderer.MantelClockBlockEntityRenderer;
 import com.chefmooon.colourfulclocks.common.block.MantelClockBlock;
 import com.chefmooon.colourfulclocks.common.block.entity.MantelClockBlockEntity;
+import com.chefmooon.colourfulclocks.common.data.types.BornholmTopGlassTypes;
 import com.chefmooon.colourfulclocks.common.data.types.PocketWatchTypes;
 import com.chefmooon.colourfulclocks.common.util.ColourfulClocksTypeUtil;
 import com.chefmooon.colourfulclocks.common.util.TextUtil;
@@ -34,9 +36,10 @@ public class MantelClockBlockEntityRendererImpl extends MantelClockBlockEntityRe
 
         PocketWatchTypes pocketWatchType = blockEntity.getDialData().pocketWatchType();
         if (pocketWatchType != PocketWatchTypes.EMPTY) {
+            poseStack.pushPose();
             renderMinuteHand(poseStack, partialTick, state);
 
-            ModelResourceLocation minuteHandLocation = new ModelResourceLocation(TextUtil.res("item/%s_minute_hand_small".formatted(BuiltInRegistries.ITEM.getKey(ColourfulClocksTypeUtil.getPocketWatchItemFromType(pocketWatchType)).getPath())), "standalone");
+            ModelResourceLocation minuteHandLocation = new ModelResourceLocation(TextUtil.res(ColourfulClocksModels.MINUTE_HAND_SMALL_PATH.formatted(BuiltInRegistries.ITEM.getKey(ColourfulClocksTypeUtil.getPocketWatchItemFromType(pocketWatchType)).getPath())), "standalone");
             BakedModel minuteHandModel = minecraft.getModelManager().getModel(minuteHandLocation);
             minecraft.getBlockRenderer().getModelRenderer().renderModel(
                     poseStack.last(),
@@ -48,7 +51,7 @@ public class MantelClockBlockEntityRendererImpl extends MantelClockBlockEntityRe
 
             renderHourHand(poseStack, partialTick, state);
 
-            ModelResourceLocation hourHandLocation = new ModelResourceLocation(TextUtil.res("item/%s_hour_hand_small".formatted(BuiltInRegistries.ITEM.getKey(ColourfulClocksTypeUtil.getPocketWatchItemFromType(pocketWatchType)).getPath())), "standalone");
+            ModelResourceLocation hourHandLocation = new ModelResourceLocation(TextUtil.res(ColourfulClocksModels.HOUR_HAND_SMALL_PATH.formatted(BuiltInRegistries.ITEM.getKey(ColourfulClocksTypeUtil.getPocketWatchItemFromType(pocketWatchType)).getPath())), "standalone");
             BakedModel hourHandModel = minecraft.getModelManager().getModel(hourHandLocation);
             minecraft.getBlockRenderer().getModelRenderer().renderModel(
                     poseStack.last(),
@@ -57,6 +60,21 @@ public class MantelClockBlockEntityRendererImpl extends MantelClockBlockEntityRe
                     hourHandModel,
                     1f, 1f, 1f,
                     packedLight, packedOverlay);
+            poseStack.popPose();
         }
+
+        poseStack.pushPose();
+        poseStack.rotateAround(getRotation(state.getValue(MantelClockBlock.FACING)), 0.5F, 0.5F, 0.5F);
+        poseStack.translate(0, 0, 0.1875F);
+        BornholmTopGlassTypes glassType = blockEntity.getDialData().getGlassType();
+        BakedModel glassModel = minecraft.getModelManager().getModel(new ModelResourceLocation(TextUtil.res(ColourfulClocksModels.DIAL_SMALL_PATH.formatted(glassType.getName())), "standalone"));
+        minecraft.getBlockRenderer().getModelRenderer().renderModel(
+                poseStack.last(),
+                bufferSource.getBuffer(RenderType.translucent()),
+                blockEntity.getBlockState(),
+                glassModel,
+                1f, 1f, 1f,
+                packedLight, packedOverlay);
+        poseStack.popPose();
     }
 }